[not verified] move i18n dropdown control to components/ folder
