+++ projects/plugins/jetpack/extensions/blocks/ai-assistant/global.d.ts
@@ -49,6 +49,7 @@ declare global {
 		};
 		Jetpack_Editor_Initial_State: {
 			adminUrl: string;
+			siteLocale: string;
 			'ai-assistant': SiteAIAssistantFeatureEndpointResponseProps;
 		};
 	}
